Moved cvr_lookup to misc module
